renaming benchmark scripts
